refactor(markdown): extract safe link components
